disabled check for localip address on diagnostics
Comment out local IP address security check

The code that verifies if the remote IP address is one of the local addresses has been commented out. This includes the array `localAddresses` and the conditional check that returns a `NotFound` result if the remote IP address is not in the list of local addresses. As a result, the application will no longer restrict access based on the remote IP address and will proceed to set the `View` property and return the page without performing this check.